add ElementAspects file
