+++ core/src/test/JSONConnector/dmos/Elements.ts
@@ -3,9 +3,8 @@
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
 import { PrimitiveType, primitiveTypeToString } from "@itwin/ecschema-metadata";
-import { GroupInformationElement, PhysicalElement, ElementUniqueAspect, PhysicalType } from "@itwin/core-backend";
-import { RelatedElementProps } from "@itwin/core-common";
-import { IRInstance, ElementDMO, PConnector, ElementAspectDMO } from "../../../pcf";
+import { GroupInformationElement, PhysicalElement, PhysicalType } from "@itwin/core-backend";
+import { IRInstance, ElementDMO, PConnector } from "../../../pcf";
 
 export const ExtPhysicalElement: ElementDMO = {
   irEntity: "ExtPhysicalElement",
@@ -50,30 +49,6 @@ export const ExtGroupInformationElement: ElementDMO = {
   ecElement: {
     name: "ExtGroupInformationElement",
     baseClass: GroupInformationElement.classFullName,
-  },
-};
-
-export const ExtElementAspect: ElementAspectDMO = {
-  irEntity: "ExtElementAspect",
-  ecElementAspect: {
-    name: "ExtElementAspect",
-    baseClass: ElementUniqueAspect.classFullName,
-    properties: [
-      {
-        name: "Name",
-        type: primitiveTypeToString(PrimitiveType.String),
-      },
-      {
-        name: "Type",
-        type: primitiveTypeToString(PrimitiveType.String),
-      },
-    ],
-  },
-  modifyProps(pc: PConnector, props: any, instance: IRInstance) {
-    props.name = instance.get("Name");
-    props.type = instance.get("Type");
-    // required
-    props.element = { id: instance.get("ExistingElementId") } as RelatedElementProps;
   },
 };
